Implement foundational data types for boolean logic
I was (and still am very torn about these).
I figure it's reasonable to try to wrap these in their own classes
(right now only `Literal` and `Clause` are) so that if I realize later
that some sort of change to the underlying data structures is reasonable
I only have to change the few methods i made to interface with the
datatype and not change many different places in the code.

On the other hand there is a very nice simplicity to simply having them
as aliases (like `using Atom = unsigned int`). I was very tempted to put
    `using Literal = int`
but I was also a bit worried I'd possibly run into issues with the fact
that then I could have 0..(4.2bn) atoms but (-2.1bn)..(2.1bn) literals,
making for a weird mismatch (although this should not matter,
practically, sincce there's no way I could ever handle nearly that many
variables).

I figure only way to find out is to try this (and possibly end up later
painstakingly changing it...)