Split up boolean-logic into three headers for each type
I've (sort of) decided on _not_ going for the simplistic approach where
a literal is just an `int` and a clause is just a `std::set<int>` since
I ran into some minor issues with it.
I figure this way might be more effort and a bit more convoluted (with a
bit of memory overhead) but it gives me more control and clearly I don't
have the capacity to anticipate what issues might come from what.