Create a barebones proof of concept for DPLL
This was the best I was able to do, as I did not manage to actually
implement it the way I envisioned a few days ago and ended up pretty
much throwing it all out today, starting "over".

This is awful in terms of performance since I pass everything by value,
have this written partly in some pseudo-functional style where I give a
set to a function and the use the set I get back from it, making for
multiple duplications in the process. (I got very scared of pointers
after my struggles and just wanted to get something working, but I dont
think my troubles were really related to that...)
It also loops over things very liberally and could probably optimized
just by being a bit clever about loops.

As a lesson-learned:
 * I'll have to really read up on when to use unique_ptr and when
   shared_ptr, because that's part of the uncertainty that tripped me up
 * I've learned a lot about maps, sets and what you can and can't do but
   I need to remember something:
     When having  doubts about whether something works the way I think
     it will, I need to just quickly try it out somewhere and make sure.
     Once I got done implementing my vision, I ended up with
     segfaulting, but because I had so many sources of uncertainty in my
     code that the task of debugging was too overwhelming for me to have
     any motivation (so I just dumpstered the whole thing)
 * I need to find a good way to test atomically